DAO: check always if the user is signed

+++ app/src/main/java/com/peppe289/echotrail/dao/user/FriendsDAO.java
@@ -5,6 +5,7 @@ import com.google.android.gms.tasks.Tasks;
 import com.google.firebase.firestore.*;
 import com.peppe289.echotrail.annotations.TestOnly;
 import com.peppe289.echotrail.controller.callback.FriendsCallback;
+import com.peppe289.echotrail.controller.user.UserController;
 import com.peppe289.echotrail.exceptions.FriendCollectionException;
 import com.peppe289.echotrail.exceptions.UserCollectionException;
 import com.peppe289.echotrail.utils.FirestoreConstants;
@@ -43,6 +44,9 @@ public class FriendsDAO {
      * @param callback The callback to be invoked upon completion.
      */
     public void requestToBeFriends(String friendId, FriendsCallback<Void, Exception> callback) {
+        if (!UserController.isLoggedIn())
+            return;
+
         String currentUserId = userDAO.getUid();
         String friendshipId = currentUserId + "_" + friendId;
 
@@ -67,6 +71,9 @@ public class FriendsDAO {
      * @param callback The callback to be invoked upon completion.
      */
     public void acceptRequest(String friendID, FriendsCallback<Void, Exception> callback) {
+        if (!UserController.isLoggedIn())
+            return;
+
         WriteBatch batch = db.batch();
 
         DocumentReference requestRef = db.collection(FirestoreConstants.COLLECTION_FRIENDS).document(userDAO.getUid() + "_" + friendID);
@@ -87,6 +94,9 @@ public class FriendsDAO {
      * @param callback The callback to be invoked upon completion.
      */
     public void rejectRequest(String friendID, FriendsCallback<String, Exception> callback) {
+        if (!UserController.isLoggedIn())
+            return;
+
         db.collection(FirestoreConstants.COLLECTION_FRIENDS)
                 .document(friendID + "_" + userDAO.getUid())
                 .delete()
@@ -101,6 +111,9 @@ public class FriendsDAO {
      * @param callback The callback to be invoked upon completion.
      */
     public void removeFriend(String pointerString, FriendsCallback<String, Exception> callback) {
+        if (!UserController.isLoggedIn())
+            return;
+
         String uid = userDAO.getUid();
 
         db.collection(FirestoreConstants.COLLECTION_USERS).document(uid)
@@ -110,6 +123,9 @@ public class FriendsDAO {
     }
 
     public void searchPendingRequests(FriendsCallback<List<String>, Exception> callback) {
+        if (!UserController.isLoggedIn())
+            return;
+
         db.collection(FirestoreConstants.COLLECTION_FRIENDS)
                 .get()
                 .addOnSuccessListener(querySnapshot -> {
@@ -139,6 +155,9 @@ public class FriendsDAO {
      * @param callback The callback to be invoked upon completion.
      */
     public void synchronizeFriendsList(FriendsCallback<Void, Exception> callback) {
+        if (!UserController.isLoggedIn())
+            return;
+
         String myUid = userDAO.getUid();
         Set<String> ignoreList = new HashSet<>();
         List<Task<Void>> pendingTasks = new ArrayList<>();
@@ -286,6 +305,9 @@ public class FriendsDAO {
      * @param callback The callback to be invoked upon completion.
      */
     public void getUIDFriendsList(String userID, FriendsCallback<List<String>, Exception> callback) {
+        if (!UserController.isLoggedIn())
+            return;
+
         db.collection(FirestoreConstants.COLLECTION_USERS)
                 .document(userID)
                 .get()
